# frontend/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Quant Platform - Quant-Service [Development]",
    "description": "Data Service API",
    "version": "v1"
  },
  "paths": {
...
    "/api/v1/market-data/economic/gdp": {
      "get": {
        "tags": [
          "Market Data v2",
          "market-data",
          "economic"
        ],
        "summary": "GDP 데이터 조회",
        "description": "미국 GDP 데이터를 조회합니다.",
        "operationId": "Market Data v2-get_gdp_data",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "데이터 간격 (annual, quarterly)",
              "default": "annual",
              "title": "Interval"
            },
            "description": "데이터 간격 (annual, quarterly)"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Market Data V2-Get Gdp Data"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market-data/economic/inflation": {
      "get": {
        "tags": [
          "Market Data v2",
          "market-data",
          "economic"
        ],
        "summary": "인플레이션 데이터 조회",
        "description": "미국 인플레이션 지표 데이터를 조회합니다.",
        "operationId": "Market Data v2-get_inflation_data",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "데이터 간격 (monthly, annual)",
              "default": "monthly",
              "title": "Interval"
            },
            "description": "데이터 간격 (monthly, annual)"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Market Data V2-Get Inflation Data"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market-data/economic/interest-rates": {
      "get": {
        "tags": [
          "Market Data v2",
          "market-data",
          "economic"
        ],
        "summary": "금리 데이터 조회",
        "description": "미국 기준금리 및 채권 수익률 데이터를 조회합니다.",
        "operationId": "Market Data v2-get_interest_rates",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "maturity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "만기 (3month, 2year, 5year, 10year, 30year)",
              "default": "10year",
              "title": "Maturity"
            },
            "description": "만기 (3month, 2year, 5year, 10year, 30year)"
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Market Data V2-Get Interest Rates"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market-data/economic/employment": {
      "get": {
        "tags": [
          "Market Data v2",
          "market-data",
          "economic"
        ],
        "summary": "고용 지표 조회",
        "description": "미국 실업률 및 고용 관련 지표를 조회합니다.",
        "operationId": "Market Data v2-get_employment_data",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Market Data V2-Get Employment Data"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/market-data/economic/consumer-sentiment": {
      "get": {
        "tags": [
          "Market Data v2",
          "market-data",
          "economic"
        ],
        "summary": "소비자 심리 지수 조회",
        "description": "미국 소비자 심리 지수를 조회합니다.",
        "operationId": "Market Data v2-get_consumer_sentiment",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Market Data V2-Get Consumer Sentiment"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}